Create README.md
airport security officials has confiscated several item of the passengers at the security check points.
all the itmes have been dumped into a huge box. each item possesses a certain amount of risk [0,1,2].
here, the risk severity of the items represent an array [] of N number of integer values. the task here is to 
sort the items based on their levels of risk in the array. the risk values range from 0 to 2.